app/requirements.txt
google-generativeai==0.8.3
httpx==0.27.2
matplotlib==3.9.2
pandas==2.2.2
qdrant-client==1.12.1
streamlit==1.37.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.5.1